Addressing PR comments and adding default integtest.sh file
