+++ bundle-workflow/src/manifests/build_manifest.py
@@ -4,6 +4,9 @@
 # this file be licensed under the Apache-2.0 license or a
 # compatible open source license.
 
+import os
+
+from aws.s3_bucket import S3Bucket
 from manifests.manifest import Manifest
 
 """
@@ -53,6 +56,18 @@ class BuildManifest(Manifest):
                 map(lambda component: component.to_dict(), self.components)
             ),
         }
+
+    @staticmethod
+    def get_build_manifest_relative_location(build_id, opensearch_version, architecture):
+        return f"builds/{opensearch_version}/{build_id}/{architecture}/manifest.yml"
+
+    @staticmethod
+    def from_s3(bucket_name, build_id, opensearch_version, architecture):
+        local_path = str(os.getcwd())
+        manifest_s3_path = BuildManifest.get_build_manifest_relative_location(build_id, opensearch_version, architecture)
+        S3Bucket(bucket_name).download_file(manifest_s3_path, local_path)
+        with open('manifest.yml', 'r') as file:
+            return BuildManifest.from_file(file)
 
     class Build:
         def __init__(self, data):
