+++ bundle-workflow/src/manifests/bundle_manifest.py
@@ -4,6 +4,9 @@
 # this file be licensed under the Apache-2.0 license or a
 # compatible open source license.
 
+import os
+
+from aws.s3_bucket import S3Bucket
 from manifests.manifest import Manifest
 
 
@@ -44,6 +47,26 @@ class BundleManifest(Manifest):
                 map(lambda component: component.to_dict(), self.components)
             ),
         }
+
+    @staticmethod
+    def from_s3(bucket_name, build_id, opensearch_version, architecture):
+        local_path = str(os.getcwd())
+        manifest_s3_path = BundleManifest.get_bundle_manifest_relative_location(build_id, opensearch_version, architecture)
+        S3Bucket(bucket_name).download_file(manifest_s3_path, local_path)
+        with open('manifest.yml', 'r') as file:
+            return BundleManifest.from_file(file)
+
+    @staticmethod
+    def get_tarball_relative_location(build_id, opensearch_version, architecture):
+        return f"bundles/{opensearch_version}/{build_id}/{architecture}/opensearch-{opensearch_version}-linux-{architecture}.tar.gz"
+
+    @staticmethod
+    def get_tarball_name(opensearch_version, architecture):
+        return f"opensearch-{opensearch_version}-linux-{architecture}.tar.gz"
+
+    @staticmethod
+    def get_bundle_manifest_relative_location(build_id, opensearch_version, architecture):
+        return f"bundles/{opensearch_version}/{build_id}/{architecture}/manifest.yml"
 
     class Build:
         def __init__(self, data):
