+++ bundle-workflow/src/paths/script_finder.py
@@ -64,12 +64,12 @@ class ScriptFinder:
     @classmethod
     def find_integ_test_script(cls, component_name, git_dir):
         paths = [
+            os.path.realpath(os.path.join(cls.default_scripts_path, "integtest.sh")),
             os.path.realpath(os.path.join(git_dir, "integtest.sh")),
             os.path.realpath(os.path.join(git_dir, "scripts/integtest.sh")),
             os.path.realpath(
                 os.path.join(cls.component_scripts_path, component_name, "integtest.sh")
             ),
-            os.path.realpath(os.path.join(cls.default_scripts_path, "integtest.sh")),
         ]
 
         return cls.__find_script("integtest.sh", paths)
